Updated Deck Class to be able to generate but exclude a set of cards specified by the dev. Jokers are still excluded by default. The Shoe class has also been updated accordingly.
Additional logic has been added to Program.cs to allow the user to select a game of Spanish 21--which excludes the 10 in all 4 suits--and was used to test the new code. It was successful. The new menu selection is commented out for now though since a new Controller will be needed to handle Spanish 21's payout structure.